+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- removed len and deref from LoadCollection, now its count -> u32
 
 ## [0.9.10] - 2025-08-07
 - removed second generic arg on load and delete many()